Marginally Better AI
AI chases the closest thing smaller than it
If nothing is smaller it runs away from things close to it that are
bigger

+++ core/src/net/avicus/battleblobs/entity/Blob.java
@@ -167,7 +167,7 @@ public class Blob extends Entity {
             double dist = distance(blob);
 
 
-            if (dist < radius) {
+            if (dist < radius && this.radius() > blob.radius()) {
                 Vector2 velocity = center.getLinearVelocity();
 
                 blob.destroy();
